+++ changelog.md
@@ -2,6 +2,7 @@
 
 ### 3.1.0 (WIP)
 
+- Inject connection object into server holded functions.
 - Show accurate position of error source.
 
 ### 3.0.0 (2026-07-07)